backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
aiofiles==25.1.0
fastapi-azure-auth==5.2.0
python-dotenv==1.2.2
pydantic==2.13.3
requests==2.33.1
psutil==7.2.2
pymongo==4.10.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-exporter-otlp==1.30.0
rich==15.0.0
debugpy==1.8.20
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
pytest-cov==7.1.0
websockets==16.0
python-jose[cryptography]==3.5.0
tinydb==4.8.2
azure-identity==1.25.3
azure-identity-broker==1.3.0; sys_platform == "win32" or sys_platform == "darwin"